fix circular loading
